Expose getAppAPI's SDK on self as global
In the worker, each app gets its own instance of the SDK.
If you want to use the sdk locally, you can either create your own instance of the sdk (which is a little complicated) or just use one of the existing ones. To do that, you expose it on `self` (which is the same as `window`, except web workers don't have `window`)
And then to use the SDK from the console you can run `self.sdk.someNamespace.someMethod()` (e.g. `self.sdk.editor.openModalPanel('token', {url: '...'})`